requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.2
Flask-SocketIO==5.3.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
six==1.16.0
urllib3==1.26.14
websocket-client==1.5.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.12.0
